restuctured functions into poker.py
