+++ .gitignore
@@ -1,3 +1,5 @@
+.gitignore
+brain-storm.md
 test.py
 # Byte-compiled / optimized / DLL files
 __pycache__/
